moved unit tests for api version
closes #122
