+++ OLT.Core.sln
@@ -211,6 +211,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "OLT.AspNetCore.Versioning",
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OLT.AspNetCore.Versioning", "src\OLT.AspNetCore.Versioning\OLT.AspNetCore.Versioning.csproj", "{828461F3-7249-461E-A72D-DAFECFD555E3}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OLT.AspNetCore.Versioning.Tests", "tests\OLT.AspNetCore.Versioning.Tests\OLT.AspNetCore.Versioning.Tests.csproj", "{F553B28A-C4AF-4215-9535-053132393CDE}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -497,6 +499,10 @@ Global
 		{828461F3-7249-461E-A72D-DAFECFD555E3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{828461F3-7249-461E-A72D-DAFECFD555E3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{828461F3-7249-461E-A72D-DAFECFD555E3}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F553B28A-C4AF-4215-9535-053132393CDE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F553B28A-C4AF-4215-9535-053132393CDE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F553B28A-C4AF-4215-9535-053132393CDE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F553B28A-C4AF-4215-9535-053132393CDE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -588,6 +594,7 @@ Global
 		{391B2EDF-30C7-44C1-AAC8-582F45FDCD6C} = {79FEE994-9D2D-43C1-A223-11078DB0F617}
 		{3E11448E-5ACC-46B8-AD23-3DD96C7F94ED} = {FC12AA4E-E1A3-4893-95F0-2230525468F2}
 		{828461F3-7249-461E-A72D-DAFECFD555E3} = {3E11448E-5ACC-46B8-AD23-3DD96C7F94ED}
+		{F553B28A-C4AF-4215-9535-053132393CDE} = {3E11448E-5ACC-46B8-AD23-3DD96C7F94ED}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {94B70EDC-6E19-4295-BFA0-9474A97DF56F}
