+++ tests/OLT.AspNetCore.Tests/Assets/TestController.cs
@@ -1,5 +1,4 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using Asp.Versioning;
 using OLT.Core;
 
 namespace OLT.AspNetCore.Tests.Assets
@@ -51,26 +50,5 @@ namespace OLT.AspNetCore.Tests.Assets
         }
     }
 
-    [ApiController]
-    [Produces("application/json")]
-    [Route("/api/api-version")]
-    public class TestVersionController : OltApiControllerBase
-    {
-       
-        [ApiVersion("1.0")]
-        [ApiVersion("2.0")]
-        [HttpGet, Route("one")]
-        public ActionResult ApiVersionOne()
-        {
-            return Ok(new { id = Faker.RandomNumber.Next() });
-        }
-
-        [ApiVersion("2.0")]
-        [HttpGet, Route("two")]
-        public ActionResult ApiVersionTwo()
-        {
-            return Ok(new { id = Faker.RandomNumber.Next() });
-        }
-
-    }
+  
 }
